+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
@@ -64,8 +63,4 @@ public class UserController {
         return userService.getMutualFriends(id, otherId);
     }
 
-    @GetMapping(value = "/users/{id}/recommendations")
-    public List<Film> getRecommendations(@PathVariable int id) {
-        return userService.getRecommendations(id);
-    }
 }
