+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
@@ -57,7 +56,4 @@ public class UserService {
         userStorage.removeUserById(userId);
     }
 
-    public List<Film> getRecommendations(int userId) {
-        return userStorage.getRecommendations(userId);
-    }
 }
