+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -31,4 +31,6 @@ public interface FilmStorage {
 
     void removeFilmById(int filmId);
 
+    List<Film> getRecommendations(int userId);
+
 }
