+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.storage.user;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.ObjectNotFoundException;
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.ArrayList;
@@ -95,10 +94,5 @@ public class InMemoryUserStorage implements UserStorage {
         } else {
             throw new ObjectNotFoundException(String.format("Пользователь не найден! Id=%d", userId));
         }
-    }
-
-    @Override
-    public List<Film> getRecommendations(int userId) {
-        return null;
     }
 }
