+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,6 +1,5 @@
 package ru.yandex.practicum.filmorate.storage.user;
 
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -24,5 +23,4 @@ public interface UserStorage {
 
     void removeUserById(int userId);
 
-    List<Film> getRecommendations(int userId);
 }
